NEW 'getTermTaxonomy' method and property to find out term's taxonomy name
Reason: UAM table stores only term's ID for locked categories and to support other taxonomies we need to know it's taxonomy name.

Solution: Run DB query (on UserGroup's init query) to find taxonomy names of all 'categories' assigned to the Group.
To support 'subcategories' find also the term's children.
Store the results in cache and a variable for quick access.